修复长度问题

+++ app/src/main/java/com/qwe7002/telegram_sms_compat/polling_json.java
@@ -1,6 +1,6 @@
 package com.qwe7002.telegram_sms_compat;
 
 class polling_json {
-    int offset;
+    long offset;
     int timeout;
 }
